Module amazon_sp_services::apis::service_api

Functions

add_appointment_for_service_job_by_service_job_id

Adds an appointment to the service job indicated by the service job identifier specified. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

assign_appointment_resources

Assigns new resource(s) or overwrite/update the existing one(s) to a service job appointment. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 1 | 2 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

cancel_reservation

Cancel a reservation. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

cancel_service_job_by_service_job_id

Cancels the service job indicated by the service job identifier specified. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

complete_service_job_by_service_job_id

Completes the service job indicated by the service job identifier specified. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

create_reservation

Create a reservation. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

create_service_document_upload_destination

Creates an upload destination. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_appointment_slots

Gets appointment slots as per the service context specified. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 20 | 40 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_appointmment_slots_by_job_id

Gets appointment slots for the service associated with the service job id specified. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_fixed_slot_capacity

Provides capacity in fixed-size slots. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_range_slot_capacity

Provides capacity slots in a format similar to availability records. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_service_job_by_service_job_id

Gets details of service job indicated by the provided serviceJobID. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 20 | 40 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

get_service_jobs

Gets service job details for the specified filter query. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 10 | 40 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

reschedule_appointment_for_service_job_by_service_job_id

Reschedules an appointment for the service job indicated by the service job identifier specified. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

set_appointment_fulfillment_data

Updates the appointment fulfillment data related to a given jobID and appointmentID. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

update_reservation

Update a reservation. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.

update_schedule

Update the schedule of the given resource. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.