Function amazon_sp_services::apis::service_api::create_service_document_upload_destination

pub async fn create_service_document_upload_destination(
    configuration: &Configuration,
    body: ServiceUploadDocument
) -> Result<CreateServiceDocumentUploadDestination, Error<CreateServiceDocumentUploadDestinationError>>

Creates an upload destination. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 5 | 20 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.