Function amazon_sp_services::apis::service_api::get_service_jobs

pub async fn get_service_jobs(
    configuration: &Configuration,
    marketplace_ids: Vec<String>,
    service_order_ids: Option<Vec<String>>,
    service_job_status: Option<Vec<String>>,
    page_token: Option<&str>,
    page_size: Option<i32>,
    sort_field: Option<&str>,
    sort_order: Option<&str>,
    created_after: Option<&str>,
    created_before: Option<&str>,
    last_updated_after: Option<&str>,
    last_updated_before: Option<&str>,
    schedule_start_date: Option<&str>,
    schedule_end_date: Option<&str>,
    asins: Option<Vec<String>>,
    required_skills: Option<Vec<String>>,
    store_ids: Option<Vec<String>>
) -> Result<GetServiceJobsResponse, Error<GetServiceJobsError>>

Gets service job details for the specified filter query. Usage Plan: | Rate (requests per second) | Burst | | –– | –– | | 10 | 40 | The x-amzn-RateLimit-Limit response header returns the usage plan rate limits that were applied to the requested operation, when available. The table above indicates the default rate and burst values for this operation. Selling partners whose business demands require higher throughput may see higher rate and burst values than those shown here. For more information, see Usage Plans and Rate Limits in the Selling Partner API.